Fix
Logout was fixed. Connected to logout with firebase to face no crash on
logout and make sure login is clear.

Added a few rules to coredata to filtering messages and chats of
specific logged in user.

Added a fix to experience as they were not updating in real time in
last commit.

Removed a few warnings.